Add Blogs.vue and FastPost.vue
