+++ src/components.d.ts
@@ -11,6 +11,7 @@ declare module 'vue' {
     Avatar: typeof import('./components/user/Avatar.vue')['default']
     AvatarCropper: typeof import('./components/user/AvatarCropper.vue')['default']
     AvatarInfo: typeof import('./components/user/AvatarInfo.vue')['default']
+    Blogs: typeof import('./components/blog/Blogs.vue')['default']
     CodeEditor: typeof import('./components/judge/CodeEditor.vue')['default']
     CodeRun: typeof import('./components/judge/CodeRun.vue')['default']
     DataMake: typeof import('./components/dakemake/DataMake.vue')['default']
@@ -60,6 +61,7 @@ declare module 'vue' {
     ElTimePicker: typeof import('element-plus/es')['ElTimePicker']
     ElTooltip: typeof import('element-plus/es')['ElTooltip']
     EmptyPage: typeof import('./components/misc/EmptyPage.vue')['default']
+    FastPost: typeof import('./components/blog/FastPost.vue')['default']
     FooterBar: typeof import('./components/FooterBar.vue')['default']
     FPSImport: typeof import('./components/problem/FPSImport.vue')['default']
     HeaderNav: typeof import('./components/HeaderNav.vue')['default']
